[RFX] Rename files for closing stage to additional information
